bdns_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"



[project]
name = "bdns-core"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "asyncpg>=0.29",
    "pydantic>=2.0",
    "python-dotenv>=1.2",
    "uuid-utils>=0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["bdns_core*"]

[tool.setuptools.package-data]
bdns_core = ["py.typed"]
